requirements.txt
absl-py
alembic
annotated-types
APScheduler
astunparse
blinker
blis
cachelib
catalogue
certifi
charset-normalizer
click
cloudpathlib
colorama
confection
config
curated-tokenizers
curated-transformers
cymem
python-dotenv

en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl#sha256=272a31e9d8530d1e075351d30a462d7e80e31da23574f1b274e200f3fff35bf5
filelock
Flask
Flask-Caching
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
flatbuffers
fsspec
gast
geographiclib
geopy
google-pasta
googlemaps
greenlet
grpcio
gunicorn
h5py
idna
itsdangerous
Jinja2
keras
langcodes
language_data
libclang
Mako
marisa-trie
Markdown
markdown-it-py
MarkupSafe
mdurl
ml-dtypes
mpmath
murmurhash
namex
networkx
numpy
opencv-python
opt_einsum
optree
packaging
preshed
protobuf
psycopg2-binary
pydantic
pydantic_core
Pygments
regex
requests
rich
shellingham
six
smart-open
spacy
spacy-curated-transformers
spacy-legacy
spacy-loggers
SQLAlchemy
srsly
sympy
tensorboard
tensorboard-data-server
tensorflow
tensorflow-io-gcs-filesystem
tensorflow-intel
termcolor
thinc
torch
tqdm
typer
typing_extensions
tzdata
tzlocal
urllib3
wasabi
weasel
Werkzeug
wrapt
google-generativeai
